Fix chat confirmation kicking issue
